pyproject.toml
[tool.poetry]
name = "homie_bot"
version = "0.1.0"
description = ""
authors = ["Egor Dmitriev <egordmitriev2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
langchain = "^0.1.5"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"
jinja2 = "^3.1.3"
transformers = "^4.37.2"
pydantic = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
